pyproject.toml
[tool.poetry]
name = "fastapi-test-app"
version = "0.1.0"
description = ""
authors = ["Nazar Annakhalov <nazar.annahalov@yandex.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.106.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.13.1"
python-dotenv = "^1.0.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}
asyncpg = "^0.29.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
celery = "^5.3.6"
flower = "^2.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
